Refactor and rename all references of pyyamlconf to anton.
