+++ docs/source/index.rst
@@ -1,4 +1,4 @@
-pyyamlconf
+anton
 ==========
 
 .. toctree::
