+++ docs/source/installation.md
@@ -3,33 +3,33 @@
 ## Install with pip
 
 ```bash
-pip install pyyamlconf
+pip install anton
 ```
 
 ## Install from source
 
 ```bash
-pip install 'git+https://github.com/karthikrangasai/pyyamlconf.git'
+pip install 'git+https://github.com/karthikrangasai/anton.git'
 ```
 
 ## Installation directions for contributors
 
-**NOTE:** Please install `pyyamlconf` with packages for testing and building docs.
+**NOTE:** Please install `anton` with packages for testing and building docs.
 
-- Fork the [repository](https://github.com/karthikrangasai/pyyamlconf.git).
+- Fork the [repository](https://github.com/karthikrangasai/anton.git).
 - Clone it locally
 
     ```bash
-    git clone https://github.com/[your username]/pyyamlconf.git
+    git clone https://github.com/[your username]/anton.git
     ```
 
 - Change the working directory
 
     ```bash
-    cd pyyamlconf
+    cd anton
     ```
 
-- Install `pyyamlconf` in editable mode with all extra packages for development
+- Install `anton` in editable mode with all extra packages for development
 
     ```bash
     poetry install --all-extras
