+++ src/anton/__init__.py
@@ -1,0 +1,3 @@
+from anton.yaml import yaml_conf
+
+__all__ = ["yaml_conf"]
