+++ src/anton/type_match.py
@@ -5,7 +5,7 @@ import warnings
 from collections import OrderedDict
 from typing import Any, Callable, Dict, List, Type, Union
 
-from pyyamlconf.constants import CONTAINER_TYPES, PRIMITIVE_TYPES
+from anton.constants import CONTAINER_TYPES, PRIMITIVE_TYPES
 
 
 def does_primitive_type_match(value: Any, parameter_type: Type) -> bool:
