--- src/pyyamlconf/__init__.py
@@ -1,3 +1,0 @@
-from pyyamlconf.yaml import yaml_conf
-
-__all__ = ["yaml_conf"]
